requirements.txt

reflex==0.8.17
PyGithub
httpx
sqlalchemy
psycopg[binary]
alembic
python-dotenv
pydantic
pydantic[email]
pydantic-settings
python-decouple
bcrypt

